Moved LogEntries into separate header.
