+++ arangod/Replication2/Methods.h
@@ -23,6 +23,7 @@
 #pragma once
 
 #include "Replication2/ReplicatedLog/LogCommon.h"
+#include "Replication2/ReplicatedLog/LogEntries.h"
 
 #include <variant>
