+++ arangod/Replication2/ReplicatedLog/LogCommon.h
@@ -146,111 +146,7 @@ auto operator<<(std::ostream& os, LogRange const& r) -> std::ostream&;
 auto intersect(LogRange a, LogRange b) noexcept -> LogRange;
 auto to_string(LogRange const&) -> std::string;
 
-struct LogPayload {
-  explicit LogPayload(velocypack::UInt8Buffer dummy);
-
-  // Named constructors, have to make copies.
-  [[nodiscard]] static auto createFromSlice(velocypack::Slice slice)
-      -> LogPayload;
-  [[nodiscard]] static auto createFromString(std::string_view string)
-      -> LogPayload;
-
-  friend auto operator==(LogPayload const&, LogPayload const&) -> bool;
-
-  [[nodiscard]] auto byteSize() const noexcept -> std::size_t;
-  [[nodiscard]] auto slice() const noexcept -> velocypack::Slice;
-
-  // just a placeholder for now
-  velocypack::UInt8Buffer dummy;
-};
-
-auto operator==(LogPayload const&, LogPayload const&) -> bool;
-
 using ParticipantId = std::string;
-
-class PersistingLogEntry {
- public:
-  PersistingLogEntry(LogTerm, LogIndex, std::optional<LogPayload>);
-  PersistingLogEntry(TermIndexPair, std::optional<LogPayload>);
-  PersistingLogEntry(LogIndex, velocypack::Slice persisted);
-
-  [[nodiscard]] auto logTerm() const noexcept -> LogTerm;
-  [[nodiscard]] auto logIndex() const noexcept -> LogIndex;
-  [[nodiscard]] auto logPayload() const noexcept
-      -> std::optional<LogPayload> const&;
-  [[nodiscard]] auto logTermIndexPair() const noexcept -> TermIndexPair;
-  [[nodiscard]] auto approxByteSize() const noexcept -> std::size_t;
-
-  class OmitLogIndex {};
-  constexpr static auto omitLogIndex = OmitLogIndex();
-  void toVelocyPack(velocypack::Builder& builder) const;
-  void toVelocyPack(velocypack::Builder& builder, OmitLogIndex) const;
-  static auto fromVelocyPack(velocypack::Slice slice) -> PersistingLogEntry;
-
-  friend auto operator==(PersistingLogEntry const&,
-                         PersistingLogEntry const&) noexcept -> bool = default;
-
- private:
-  void entriesWithoutIndexToVelocyPack(velocypack::Builder& builder) const;
-
-  LogTerm _logTerm{};
-  LogIndex _logIndex{};
-  // TODO It seems impractical to not copy persisting log entries, so we should
-  //      probably make this a shared_ptr (or immer::box).
-  std::optional<LogPayload> _payload;
-
-  // TODO this is a magic constant "measuring" the size of
-  //      of the non-payload data in a PersistingLogEntry
-  static inline constexpr auto approxMetaDataSize = std::size_t{42};
-};
-
-// A log entry, enriched with non-persisted metadata, to be stored in an
-// InMemoryLog.
-class InMemoryLogEntry {
- public:
-  using clock = std::chrono::steady_clock;
-
-  explicit InMemoryLogEntry(PersistingLogEntry entry, bool waitForSync = false);
-
-  [[nodiscard]] auto insertTp() const noexcept -> clock::time_point;
-  void setInsertTp(clock::time_point) noexcept;
-  [[nodiscard]] auto entry() const noexcept -> PersistingLogEntry const&;
-  [[nodiscard]] bool getWaitForSync() const noexcept { return _waitForSync; }
-
- private:
-  bool _waitForSync;
-  // Immutable box that allows sharing, i.e. cheap copying.
-  ::immer::box<PersistingLogEntry, ::arangodb::immer::arango_memory_policy>
-      _logEntry;
-  // Timepoint at which the insert was started (not the point in time where it
-  // was committed)
-  clock::time_point _insertTp{};
-};
-
-// A log entry as visible to the user of a replicated log.
-// Does thus always contain a payload: only internal log entries are without
-// payload, which aren't visible to the user. User-defined log entries always
-// contain a payload.
-// The term is not of interest, and therefore not part of this struct.
-// Note that when these entries are visible, they are already committed.
-// It does not own the payload, so make sure it is still valid when using it.
-class LogEntryView {
- public:
-  LogEntryView() = delete;
-  LogEntryView(LogIndex index, LogPayload const& payload) noexcept;
-  LogEntryView(LogIndex index, velocypack::Slice payload) noexcept;
-
-  [[nodiscard]] auto logIndex() const noexcept -> LogIndex;
-  [[nodiscard]] auto logPayload() const noexcept -> velocypack::Slice;
-  [[nodiscard]] auto clonePayload() const -> LogPayload;
-
-  void toVelocyPack(velocypack::Builder& builder) const;
-  static auto fromVelocyPack(velocypack::Slice slice) -> LogEntryView;
-
- private:
-  LogIndex _index;
-  velocypack::Slice _payload;
-};
 
 class LogId : public arangodb::basics::Identifier {
  public:
@@ -262,29 +158,6 @@ class LogId : public arangodb::basics::Identifier {
 };
 
 auto to_string(LogId logId) -> std::string;
-
-template<typename T>
-struct TypedLogIterator {
-  virtual ~TypedLogIterator() = default;
-  // The returned view is guaranteed to stay valid until a successive next()
-  // call (only).
-  virtual auto next() -> std::optional<T> = 0;
-};
-
-template<typename T>
-struct TypedLogRangeIterator : TypedLogIterator<T> {
-  // returns the index interval [from, to)
-  // Note that this does not imply that all indexes in the range [from, to)
-  // are returned. Hence (to - from) is only an upper bound on the number of
-  // entries returned.
-  [[nodiscard]] virtual auto range() const noexcept -> LogRange = 0;
-};
-
-using LogIterator = TypedLogIterator<LogEntryView>;
-using LogRangeIterator = TypedLogRangeIterator<LogEntryView>;
-
-// ReplicatedLog-internal iterator over PersistingLogEntries
-struct PersistedLogIterator : TypedLogIterator<PersistingLogEntry> {};
 
 struct LogConfig {
   std::size_t writeConcern = 1;
