+++ arangod/Replication2/Streams/Streams.h
@@ -22,7 +22,8 @@
 ////////////////////////////////////////////////////////////////////////////////
 #pragma once
 
-#include <Replication2/ReplicatedLog/LogCommon.h>
+#include "Replication2/ReplicatedLog/LogCommon.h"
+#include "Replication2/ReplicatedLog/LogEntries.h"
 #include "Replication2/Streams/StreamSpecification.h"
 
 namespace arangodb::futures {
